+++ rollup.config.js
@@ -71,6 +71,9 @@ const config = [
   getConfig('./src/preact/index.ts', 'preact/index'),
   getConfig('./src/react/index.ts', 'react/index'),
   getConfig('./src/svelte/index.ts', 'svelte/index'),
+  getConfig('./src/vue/index.ts', 'vue/index'),
+  getConfig('./src/middleware/index.ts', 'middleware/index'),
+  getConfig('./src/utils/index.ts', 'utils/index'),
 ]
 
 export default config