+++ src/utils/bindActions.ts
@@ -1,17 +1,18 @@
+import set from "./set"
+
 export default function bindActions(actions, store) {
   actions = typeof actions === "function" ? actions(store) : actions
 
   let bound = {}
   for (let name in actions) {
     bound[name] = (...args) => {
-      let action = store.middlewares.reduce((newAction, middleware) => {
-        return middleware(store)(newAction, name)
-      }, actions[name])
-      let ret = action(store.getState(), ...args)
-      if (ret != null) {
-        if (ret.then) return ret.then(store.setState)
-        store.setState(ret)
+      const action = actions[name]
+
+      if (typeof store.middleware === "function") {
+        return store.middleware(store, action, args)
       }
+
+      return set(store, action(store.getState(), ...args))
     }
   }